Add: /werewolf

+++ src/commands/miscellaneous/mass-ping.ts
@@ -70,24 +70,17 @@ const command: Command = {
 		const message = interaction.options.getString("message", false);
 
 		if (number < 1 || (!message && number > 1000) || (message && number > 100))
-			return interaction.followUp(
-				translations.strings.invalid_number(Boolean(message)),
-			);
+			return interaction.followUp(translations.strings.invalid_number(Boolean(message)));
 
 		interaction.followUp(translations.strings.sending());
 
 		const messages: Promise<Message>[] = [];
 
 		for (let i = 0; i < number; i++) {
-			messages.push(
-				interaction.channel.send(`${user.toString()} ${message ?? ""}`),
-			);
+			messages.push(interaction.channel.send(`${user.toString()} ${message ?? ""}`));
 		}
 
-		if (!message)
-			await (interaction.channel as TextChannel).bulkDelete(
-				await Promise.all(messages),
-			);
+		if (!message) await (interaction.channel as TextChannel).bulkDelete(await Promise.all(messages));
 
 		interaction.editReply(translations.strings.done());
 	},
